backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "LLMGuard: Real-Time LLM Guard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "modal>=0.64.0",
    "vllm>=0.6.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "PyGithub>=1.59.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
]
